requirements.txt
fastapi
uvicorn[standard]
alembic
aiosqlite
python-jose 
pytest
bcrypt
python-decouple
httpx
pytest-asyncio